dom/bindings/test/test_bug1457051.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1457051
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1457051</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://global/skin"/>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

  /** Test for Bug 1457051 **/
  ok(Element.isInstance(document.documentElement), "Basic isInstance works");
  ok(!Element.isInstance(null),
     "Passing null should return false without throwing");
  ok(!Element.isInstance(5), "Passing 5 should return false without throwing");
  var obj = Object.create(Element.prototype);
  ok(obj instanceof Element, "instanceof should walk the proto chain");
  ok(!Element.isInstance(obj), "isInstance should be a pure brand check");

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1457051">Mozilla Bug 1457051</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
